rename ParamsChecker
